sudo main is checked and save function is modified
What changes you have made?
.....
-Changed save function.function is not changing incoming list instead it’s
 creating a copy of the list and changing it.
-I have also added test for sudo-main and is testing for save and query.

Why this change is necessary?
.....
-changing the save function was necessary to protect the
function,because previously the list was changing globaly.
-Adding tests ware necessary to check every thing is working fine or not.

What is the impact of the change?
......
the changes and addition of tests made the system more relible.Also the
tests are for sudo-main so it's testing the system from a very top
level.